+++ plugins/processors/clickdetect/clickdetect_test.go
@@ -51,13 +51,13 @@ func TestEmit0WhenClickIdEqual(t *testing.T) {
 	var result = cd.Apply([]telegraf.Metric{metric1}...)
 	var changedFieldValue, changedFieldExists = result[0].GetField(ChangedField)
 	require.True(t, changedFieldExists)
-	require.Equal(t, float64(0), changedFieldValue)
+	require.Equal(t, int64(0), changedFieldValue)
 	require.False(t, result[0].HasField(FieldToChange))
 
 	result = cd.Apply([]telegraf.Metric{metric2}...)
 	changedFieldValue, changedFieldExists = result[0].GetField(ChangedField)
 	require.True(t, changedFieldExists)
-	require.Equal(t, float64(0), changedFieldValue)
+	require.Equal(t, int64(0), changedFieldValue)
 	require.False(t, result[0].HasField(FieldToChange))
 }
 
@@ -99,13 +99,13 @@ func TestNotEmit0WhenClickIdGreater(t *testing.T) {
 	var result = cd.Apply([]telegraf.Metric{metric1}...)
 	var changedFieldValue, changedFieldExists = result[0].GetField(ChangedField)
 	require.True(t, changedFieldExists)
-	require.Equal(t, float64(0), changedFieldValue)
+	require.Equal(t, int64(0), changedFieldValue)
 	require.False(t, result[0].HasField(FieldToChange))
 
 	result = cd.Apply([]telegraf.Metric{metric2}...)
 	changedFieldValue, changedFieldExists = result[0].GetField(ChangedField)
 	require.True(t, changedFieldExists)
-	require.Equal(t, float64(6), changedFieldValue)
+	require.Equal(t, int64(6), changedFieldValue)
 	require.False(t, result[0].HasField(FieldToChange))
 }
 
@@ -147,13 +147,13 @@ func TestNotEmit0WhenClickIdRolledBack(t *testing.T) {
 	var result = cd.Apply([]telegraf.Metric{metric1}...)
 	var changedFieldValue, changedFieldExists = result[0].GetField(ChangedField)
 	require.True(t, changedFieldExists)
-	require.Equal(t, float64(0), changedFieldValue)
+	require.Equal(t, int64(0), changedFieldValue)
 	require.False(t, result[0].HasField(FieldToChange))
 
 	result = cd.Apply([]telegraf.Metric{metric2}...)
 	changedFieldValue, changedFieldExists = result[0].GetField(ChangedField)
 	require.True(t, changedFieldExists)
-	require.Equal(t, float64(8), changedFieldValue)
+	require.Equal(t, int64(8), changedFieldValue)
 	require.False(t, result[0].HasField(FieldToChange))
 }
 
@@ -195,12 +195,12 @@ func TestNotEmit0WhenClickIdLower(t *testing.T) {
 	var result = cd.Apply([]telegraf.Metric{metric1}...)
 	var changedFieldValue, changedFieldExists = result[0].GetField(ChangedField)
 	require.True(t, changedFieldExists)
-	require.Equal(t, float64(0), changedFieldValue)
+	require.Equal(t, int64(0), changedFieldValue)
 	require.False(t, result[0].HasField(FieldToChange))
 
 	result = cd.Apply([]telegraf.Metric{metric2}...)
 	changedFieldValue, changedFieldExists = result[0].GetField(ChangedField)
 	require.True(t, changedFieldExists)
-	require.Equal(t, float64(0), changedFieldValue)
+	require.Equal(t, int64(0), changedFieldValue)
 	require.False(t, result[0].HasField(FieldToChange))
 }
